+++ CHANGELOG.md
@@ -8,6 +8,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com), and this 
 ### Added
 * Add `CacheNamespaceConfigOption` to customize the cache namespace.
 * Add support for PHP 8.3
+* [#200](https://github.com/shlinkio/shlink-installer/issues/200) Add Matomo integration config options.
 
 ### Changed
 * *Nothing*